Got init working!
Tried to call and use window before it was used which was the issue of
the panic. Now I inits more wgpu stuff but there is still much left.
Also since some of the wgpu methods are async a async engine needed to
be selected. Went with futures crate for now but might switch over to
pollster later on.